+++ openpyxl2/packaging/tests/test_manifest.py
@@ -262,6 +262,15 @@ class TestContentTypes:
         diff = compare_xml(xml, expected)
         assert diff is None, diff
 
+    def test_vba(self, datadir):
+        from openpyxl2 import load_workbook
+        from ..manifest import write_content_types
+        datadir.chdir()
+        wb = load_workbook('sample.xlsm', keep_vba=True)
+        manifest = write_content_types(wb)
+        partnames = [t.PartName for t in manifest.Override]
+        expected = ['/xl/workbook.xml', '/xl/worksheets/sheet1.xml', '/xl/worksheets/sheet2.xml', '/xl/worksheets/sheet3.xml', '/xl/theme/theme1.xml', '/xl/styles.xml', '/docProps/core.xml', '/docProps/app.xml', '/xl/sharedStrings.xml']
+        assert partnames == expected
 
     @pytest.mark.lxml_required # for XPATH lookup
     @pytest.mark.parametrize("has_vba, as_template, content_type",