+++ doc/changes.rst
@@ -35,6 +35,11 @@ Bug fixes
 ==================
 
 
+Bug fixes
+---------
+
+* `#594 <https://bitbucket.org/openpyxl/openpyxl/issues/594>`_ Content types might be missing when keeping VBA
+
 
 2.3.3 (2016-01-18)
 ==================